feat: Add conditional logic to navigate to game after nap activity
The code changes in `activity.tsx` include adding conditional logic to navigate to the game page after the nap activity. If the current time is between 5 AM and 11 PM, the `napLabel` callback is called. Otherwise, the `sleepLabel` callback is called. This change improves the user experience by providing appropriate navigation based on the time of day.